Fix #13795: size_t mayn't be the same size as uint64_t.

+++ src/engine_base.h
@@ -213,7 +213,7 @@ struct EngineIDMapping {
 
 /** Projection to get a unique key of an EngineIDMapping, used for sorting in EngineOverrideManager. */
 struct EngineIDMappingKeyProjection {
-	inline size_t operator()(const EngineIDMapping &eid) const { return eid.Key(); }
+	inline uint64_t operator()(const EngineIDMapping &eid) const { return eid.Key(); }
 };
 
 /**
